Application program-7
Given an N×N chessboard, the task is to place N queens on the board
so that no two queens can attack each other. The objective of this
case study is to implement the N-Queens algorithm in C Language to
find all possible arrangements of the queens on the chessboard.